+++ forge/changelog.md
@@ -1,2 +1,1 @@
-- Updated whisper icons
-- Updated lock icon
+- Added cross compatibility between Fabric, Forge and Bukkit/Spigot/Paper